+++ src/build_yw2xtg.py
@@ -18,10 +18,10 @@ TARGET_FILE = f'{BUILD}yw2xtg.pyw'
 
 
 def main():
-    # inliner.run(SOURCE_FILE, TARGET_FILE, 'yw2xtglib', '../src/', copyPyWriter=True)
-    # inliner.run(TARGET_FILE, TARGET_FILE, 'pywriter', '../../PyWriter/src/', copyPyWriter=True)
-    inliner.run(SOURCE_FILE, TARGET_FILE, 'yw2xtglib', '../src/')
-    inliner.run(TARGET_FILE, TARGET_FILE, 'pywriter', '../src/')
+    inliner.run(SOURCE_FILE, TARGET_FILE, 'yw2xtglib', '../src/', copyPyWriter=False)
+    inliner.run(TARGET_FILE, TARGET_FILE, 'pywriter', '../../PyWriter/src/', copyPyWriter=False)
+    # inliner.run(SOURCE_FILE, TARGET_FILE, 'yw2xtglib', '../src/')
+    # inliner.run(TARGET_FILE, TARGET_FILE, 'pywriter', '../src/')
     print('Done.')